+++ src/pudl/output/pudltabl.py
@@ -163,7 +163,7 @@ class PudlTabl:
             "out_eia__yearly_plants": "plants_eia860",
             "out_eia__yearly_utilities": "utils_eia860",
             "out_eia__yearly_boilers": "boil_eia860",
-            "out_eia__yearly_generators": "gens_eia860",
+            "_out_eia__yearly_generators": "gens_eia860",
             "_out_eia__plants_utilities": "pu_eia860",
             # eia860 (denormalized, data primarily from EIA-860)
             "out_eia860__yearly_ownership": "own_eia860",
@@ -218,17 +218,21 @@ class PudlTabl:
             "out_eia861__compiled_geometry_utilities": "compiled_geometry_utility_eia861",
             # state demand
             "out_ferc714__hourly_predicted_state_demand": "predicted_state_hourly_demand",
+            # plant parts
+            "out_eia__yearly_generators_by_ownership": "gens_mega_eia",
+            "out_eia__yearly_plant_parts": "plant_parts_eia",
+            "out__yearly_plants_all_ferc1_plant_parts_eia": "ferc1_eia",
         }
 
         table_method_map_any_agg = {
             "out_eia923__AGG_generation_fuel_by_generator_energy_source": "gen_fuel_by_generator_energy_source_eia923",
             "out_eia923__AGG_generation_fuel_by_generator": "gen_fuel_by_generator_eia923",
-            "heat_rate_by_unit_AGG": "hr_by_unit",
-            "heat_rate_by_generator_AGG": "hr_by_gen",
-            "capacity_factor_by_generator_AGG": "capacity_factor",
-            "fuel_cost_by_generator_AGG": "fuel_cost",
-            "mcoe_AGG": "mcoe",
-            "mcoe_generators_AGG": "mcoe_generators",
+            "_out_eia__AGG_heat_rate_by_unit": "hr_by_unit",
+            "_out_eia__AGG_heat_rate_by_generator": "hr_by_gen",
+            "_out_eia__AGG_capacity_factor_by_generator": "capacity_factor",
+            "_out_eia__AGG_fuel_cost_by_generator": "fuel_cost",
+            "_out_eia__AGG_derived_generator_attributes": "mcoe",
+            "out_eia__AGG_generators": "mcoe_generators",
         }
 
         table_method_map_yearly_only = {
@@ -403,95 +407,8 @@ class PudlTabl:
         return gen_df
 
     ###########################################################################
-    # Plant Parts EIA outputs
-    ###########################################################################
-    def gens_mega_eia(
-        self: Self,
-        update: bool = False,
-        gens_cols: Literal["all"] | list[str] | None = None,
-    ) -> pd.DataFrame:
-        """Generate and return a generators table with ownership integrated.
-
-        Args:
-            update: If True, re-calculate dataframe even if a cached version exists.
-            gens_cols: Ignored. Retained for backwards compatibility only.
-
-        Returns:
-            A table of all of the generators with identifying columns and data columns,
-            sliced by ownership which makes "total" and "owned" records for each
-            generator owner. The "owned" records have the generator's data scaled to the
-            ownership percentage (e.g. if a 100 MW generator has a 75% stake owner and a
-            25% stake owner, this will result in two "owned" records with 75 MW and 25
-            MW). The "total" records correspond to the full plant for every owner (e.g.
-            using the same 2-owner 100 MW generator as above, each owner will have a
-            records with 100 MW).
-
-        Raises:
-            AssertionError: If the frequency of the pudl_out object is not ``AS``.
-        """
-        if update or self._dfs["gens_mega_eia"] is None:
-            if self.freq != "AS":
-                raise AssertionError(
-                    "The frequency of the pudl_out object must be `AS` for the "
-                    f"plant-parts table and we got {self.freq}"
-                )
-
-            self._dfs[
-                "gens_mega_eia"
-            ] = pudl.analysis.plant_parts_eia.MakeMegaGenTbl().execute(
-                mcoe=self.mcoe_generators(),
-                own_eia860=self.own_eia860(),
-            )
-        return self._dfs["gens_mega_eia"]
-
-    def plant_parts_eia(
-        self: Self,
-        update: bool = False,
-        update_gens_mega: bool = False,
-        gens_cols: Literal["all"] | list[str] | None = None,
-    ) -> pd.DataFrame:
-        """Generate and return master plant-parts EIA.
-
-        Args:
-            update: If True, re-calculate dataframe even if a cached version exists.
-            update_gens_mega: If True, update the gigantic Gens Mega table.
-            gens_cols: Ignored. Retained for backwards compatibility only.
-        """
-        update_any = any([update, update_gens_mega])
-        if update_any or self._dfs["plant_parts_eia"] is None:
-            # make the plant-parts objects
-            self.parts_compiler = pudl.analysis.plant_parts_eia.MakePlantParts(self)
-            # make the plant-parts df!
-            self._dfs["plant_parts_eia"] = self.parts_compiler.execute(
-                gens_mega=self.gens_mega_eia(
-                    update=update_gens_mega,
-                )
-            )
-
-        return self._dfs["plant_parts_eia"]
-
-    ###########################################################################
     # GLUE OUTPUTS
     ###########################################################################
-    def ferc1_eia(
-        self: Self,
-        update: bool = False,
-        update_plant_parts_eia: bool = False,
-        update_plants_all_ferc1: bool = False,
-        update_fbp_ferc1: bool = False,
-    ) -> pd.DataFrame:
-        """Generate the connection between FERC1 and EIA."""
-        update_any = any(
-            [update, update_plant_parts_eia, update_plants_all_ferc1, update_fbp_ferc1]
-        )
-        if update_any or self._dfs["ferc1_eia"] is None:
-            self._dfs["ferc1_eia"] = pudl.analysis.ferc1_eia.execute(
-                plant_parts_eia=self.plant_parts_eia(update=update_plant_parts_eia),
-                plants_all_ferc1=self.plants_all_ferc1(),
-                fbp_ferc1=self.fbp_ferc1(),
-            )
-        return self._dfs["ferc1_eia"]
-
     def epacamd_eia(self: Self) -> pd.DataFrame:
         """Read the EPACAMD-EIA Crosswalk from the PUDL DB."""
         return pd.read_sql("core_epa__assn_epacamd_eia", self.pudl_engine).pipe(