+++ src/pudl/transform/params/ferc1.py
@@ -299,6 +299,8 @@ UTILITY_TYPE_CATEGORIES = {
         "electric": {"electric", "ferc:ElectricUtilityMember"},
         "gas": {"gas", "ferc:GasUtilityMember"},
         "common": {"common", "ferc:CommonUtilityMember"},
+        "other": {"other", "ferc:OtherUtilityMember"},
+        "total": {"total"},
     }
 }
 
@@ -326,6 +328,7 @@ FUEL_CATEGORIES: dict[str, set[str]] = {
     "categories": {
         "coal": {
             "bit coal",
+            "bit-coal",
             "ciak",
             "coa",
             "coal",
@@ -2724,7 +2727,19 @@ TRANSFORM_PARAMS = {
             "on": "ferc_account_label",
         },
         "align_row_numbers_dbf": {"dbf_table_names": ["f1_plant_in_srvce"]},
-        "reconcile_table_calculations": {"column_to_check": "ending_balance"},
+        # Known issue with reporting of construction in progress not classified in classified fields of table.
+        "reconcile_table_calculations": {
+            "column_to_check": "ending_balance",
+            "group_metric_checks": {
+                "group_metric_tolerances": {
+                    "ungrouped": {"error_frequency": 0.08},
+                    "utility_id_ferc1": {
+                        "error_frequency": 0.038,
+                        "relative_error_magnitude": 0.11,
+                    },
+                },
+            },
+        },
     },
     "core_ferc1__yearly_plants_pumped_storage": {
         "rename_columns_ferc1": {
@@ -2911,8 +2926,8 @@ TRANSFORM_PARAMS = {
                     "energy_charges_of_purchased_power": "energy_charges",
                     "energy_delivered_through_power_exchanges": "delivered_mwh",
                     "energy_received_through_power_exchanges": "received_mwh",
-                    "megawatt_hours_purchased_for_energy_storage": "purchased_storage_mwh",  # new column
-                    "megawatt_hours_purchased_other_than_storage": "purchased_mwh",
+                    "megawatt_hours_purchased_for_energy_storage": "purchased_storage_mwh",
+                    "megawatt_hours_purchased_other_than_storage": "purchased_other_than_storage_mwh",
                     "name_of_company_or_public_authority_providing_purchased_power": "seller_name",
                     "other_charges_of_purchased_power": "other_charges",
                     "purchased_power_axis": "purchased_power_axis",  # new probably pk col
@@ -2942,6 +2957,7 @@ TRANSFORM_PARAMS = {
                     "other_charges",
                     "total_settlement",
                     "purchased_storage_mwh",
+                    "purchased_other_than_storage_mwh",
                     "billing_demand_mw",
                     "non_coincident_peak_demand_mw",
                     "coincident_peak_demand_mw",
@@ -3308,8 +3324,6 @@ TRANSFORM_PARAMS = {
             "utility_type": {
                 "categories": UTILITY_TYPE_CATEGORIES["categories"]
                 | {
-                    "total": {"total"},
-                    "other1": {"other1", "ferc:OtherUtilityMember"},
                     "other2": {"other2", "ferc:OtherUtility2Member"},
                     "other3": {"other3", "ferc:OtherUtility3Member"},
                 }
@@ -3324,6 +3338,29 @@ TRANSFORM_PARAMS = {
         "reconcile_table_calculations": {
             "column_to_check": "ending_balance",
             "subtotal_column": "utility_type",
+            "group_metric_checks": {
+                "group_metric_tolerances": {
+                    "ungrouped": {
+                        "error_frequency": 0.0092,
+                        "relative_error_magnitude": 0.039,
+                    },
+                    "report_year": {
+                        "error_frequency": 0.24,  # 2021 is bad :-/
+                        "null_calculated_value_frequency": 1.0,
+                        "relative_error_magnitude": 0.47,  # 2021 is bad :-/
+                    },
+                    "xbrl_factoid": {
+                        "error_frequency": 0.16,
+                        "relative_error_magnitude": 0.2,
+                        "null_calculated_value_frequency": 1.0,
+                    },  # utility_plant_and_construction_work_in_progress bad
+                    "utility_id_ferc1": {
+                        "error_frequency": 0.21,  # 444 is bad
+                        "relative_error_magnitude": 0.074,
+                        "null_calculated_value_frequency": 1.0,
+                    },
+                },
+            },
         },
     },
     "core_ferc1__yearly_balance_sheet_assets": {
@@ -3347,6 +3384,7 @@ TRANSFORM_PARAMS = {
                     "entity_id": "utility_id_ferc1_xbrl",
                     "report_year": "report_year",
                     "xbrl_factoid": "asset_type",
+                    "preliminary_natural_gas_survey_and_investigation_charges_and_other_preliminary_survey_and_investigation_charges": "preliminary_natural_gas_survey_and_investigation_charges",
                 }
             },
         },
@@ -3374,7 +3412,14 @@ TRANSFORM_PARAMS = {
             "rename_columns": {"xbrl_factoid": "asset_type"},
             "on": "asset_type",
         },
-        "reconcile_table_calculations": {"column_to_check": "ending_balance"},
+        "reconcile_table_calculations": {
+            "column_to_check": "ending_balance",
+            "group_metric_checks": {
+                "group_metric_tolerances": {
+                    "ungrouped": {"error_frequency": 0.0002},
+                },
+            },
+        },
     },
     "core_ferc1__yearly_balance_sheet_liabilities": {
         "rename_columns_ferc1": {
@@ -3500,6 +3545,12 @@ TRANSFORM_PARAMS = {
         "reconcile_table_calculations": {
             "column_to_check": "dollar_value",
             "subtotal_column": "plant_function",
+            "group_metric_checks": {
+                "group_metric_tolerances": {
+                    "ungrouped": {"null_calculated_value_frequency": 0.84},
+                    "report_year": {"null_calculated_value_frequency": 1.0},
+                },
+            },
         },
     },
     "core_ferc1__yearly_electric_operating_revenues": {
@@ -3631,7 +3682,23 @@ TRANSFORM_PARAMS = {
                 ],
             },
         ],
-        "reconcile_table_calculations": {"column_to_check": "dollar_value"},
+        "reconcile_table_calculations": {
+            "column_to_check": "dollar_value",
+            "group_metric_checks": {
+                "group_metric_tolerances": {
+                    "ungrouped": {"error_frequency": 0.0031},
+                    "xbrl_factoid": {
+                        "error_frequency": 0.04,
+                        "null_calculated_value_frequency": 1.0,
+                        "relative_error_magnitude": 0.03,
+                    },  # other_operating_revenues bad
+                    "utility_id_ferc1": {
+                        "error_frequency": 0.17,
+                        "null_calculated_value_frequency": 1.0,
+                    },
+                },
+            },
+        },
     },
     "core_ferc1__yearly_retained_earnings": {
         "rename_columns_ferc1": {
@@ -3709,7 +3776,9 @@ TRANSFORM_PARAMS = {
         "strip_non_numeric_values": {"amount": {"strip_non_numeric_values": True}},
         "reconcile_table_calculations": {
             "column_to_check": "ending_balance",
-            "calculation_tolerance": 0.08,
+            "group_metric_checks": {
+                "group_metric_tolerances": {"ungrouped": {"error_frequency": 0.08}},
+            },
         },
     },
     "core_ferc1__yearly_income_statement": {
@@ -3875,6 +3944,17 @@ TRANSFORM_PARAMS = {
         "reconcile_table_calculations": {
             "column_to_check": "dollar_value",
             "subtotal_column": "utility_type",
+            "group_metric_checks": {
+                "group_metric_tolerances": {
+                    "ungrouped": {"null_calculated_value_frequency": 0.74},
+                    "report_year": {"null_calculated_value_frequency": 1.0},
+                    "utility_id_ferc1": {
+                        "error_frequency": 0.042,
+                        "relative_error_magnitude": 0.22,
+                        "null_calculated_value_frequency": 1.0,
+                    },
+                },
+            },
         },
     },
     "core_ferc1__yearly_electric_plant_depreciation_changes": {
@@ -3976,8 +4056,26 @@ TRANSFORM_PARAMS = {
             "column_to_check": "dollar_value",
             # Note: this table does not currently get exploded. It will require
             # additional debugging at a later date.
-            "calculation_tolerance": 0.4,
-            "subtotal_column": "depreciation_type",
+            "group_metric_checks": {
+                "group_metric_tolerances": {
+                    "ungrouped": {
+                        "error_frequency": 0.058,
+                        "relative_error_magnitude": 0.045,
+                    },
+                    "report_year": {
+                        "error_frequency": 0.078,
+                        "relative_error_magnitude": 0.073,
+                    },
+                    "utility_id_ferc1": {
+                        "error_frequency": 0.14,
+                        "relative_error_magnitude": 0.66,
+                    },
+                    "xbrl_factoid": {
+                        "error_frequency": 0.58,
+                        "relative_error_magnitude": 0.1,
+                    },  # ending_balance is bad
+                },
+            },
         },
     },
     "core_ferc1__yearly_electric_plant_depreciation_functional": {
@@ -4053,8 +4151,8 @@ TRANSFORM_PARAMS = {
             },
         },
         "merge_xbrl_metadata": {
-            "rename_columns": {"xbrl_factoid": "plant_function"},
-            "on": "plant_function",
+            "rename_columns": {"xbrl_factoid": "depreciation_type"},
+            "on": "depreciation_type",
         },
         "select_dbf_rows_by_category": {
             "column_name": "plant_function",
@@ -4425,7 +4523,23 @@ TRANSFORM_PARAMS = {
             "rename_columns": {"xbrl_factoid": "expense_type"},
             "on": "expense_type",
         },
-        "reconcile_table_calculations": {"column_to_check": "dollar_value"},
+        "reconcile_table_calculations": {
+            "column_to_check": "dollar_value",
+            "group_metric_checks": {
+                "group_metric_tolerances": {
+                    "ungrouped": {"relative_error_magnitude": 0.002},
+                    "report_year": {"relative_error_magnitude": 0.042},
+                    "xbrl_factoid": {
+                        "error_frequency": 0.018,
+                        "relative_error_magnitude": 0.028,
+                    },
+                    "utility_id_ferc1": {
+                        "error_frequency": 0.017,
+                        "relative_error_magnitude": 0.066,
+                    },
+                },
+            },
+        },
     },
     "core_ferc1__yearly_other_regulatory_liabilities": {
         "rename_columns_ferc1": {