+++ docs/release_notes.rst
@@ -32,12 +32,13 @@ Dagster Adoption
     returns the resources ids for a given etl group.
   * :mod:`pudl.settings.FercToSqliteSettings` class now loads all FERC
     datasources if no datasets are specified.
+  * The Excel extractor in ``pudl.extract.excel`` has been updated to parallelize
+    Excel spreadsheet extraction using Dagster ``@multi_asset`` functionality, thanks to
+    :user:`dstansby`. This is currently being used for EIA 860, 861 and 923 data. See
+    :issue:`2385` and PRs :pr:`2644`, :pr:`2943`.
 
 * EIA ETL changes:
 
-  * EIA extract methods are now ``@multi_asset`` that return an asset for each
-    raw table. 860 and 923 are separate ``@multi_asset`` which allows this data
-    to be extracted in parallel.
   * The EIA table level cleaning functions are now
     dagster assets. The table level cleaning assets now have a "clean\_" prefix
     and a "_{datasource}" suffix to distinguish them from the final harvested tables.
@@ -109,6 +110,12 @@ Data Coverage
   * :ref:`core_ferc1__yearly_cash_flow`, see issue :issue:`1821` & PR :pr:`2184`
   * :ref:`core_ferc1__yearly_electricity_sales_by_rate_schedule`, see issue
     :issue:`1823` & PR :pr:`2205`.
+
+* Harvested owner utilities from the EIA 860 ownership table which are now included in
+  the :ref:`core_eia__entity_utilities` and :ref:`core_pudl__assn_utilities_eia`
+  tables. See :pr:`2714`. Renamed columns with owner or operator suffix to differentiate
+  between owner and operator utility columns in :ref:`core_eia860__scd_ownership` and
+  :ref:`out_eia860__yearly_ownership`. See :pr:`2903`.
 
 * New PUDL tables from :doc:`data_sources/eia860`:
 
@@ -182,6 +189,10 @@ Data Coverage
   and uses these complete IDs to develop a full-coverage ``subplant_id`` column which
   granularly connects EPA CAMD with EIA. Thanks to :user:`grgmiller` for his
   contribution to this process. See :issue:`2456` & :pr:`2491`.
+
+* Added new table :ref:`out__yearly_plants_all_ferc1_plant_parts_eia` which links FERC1
+  records from :ref:`out_ferc1__yearly_all_plants` and
+  :ref:`out_eia__yearly_plant_parts`.
 
 * Thanks to contributions from :user:`rousik` we've generalized the code we use to
   convert FERC's old annual Visual FoxPro databases into multi-year SQLite databases.